🚚 (keys-das): Rename service dir to services
